% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_tree_discovery.R
\name{add_clustering_single_cell_dataset}
\alias{add_clustering_single_cell_dataset}
\title{Add Clustering Information to Single Cell Dataset}
\usage{
add_clustering_single_cell_dataset(
  sc.eset,
  var_cluster_name = "cluster_normalized",
  fit_hierarchical
)
}
\arguments{
\item{sc.eset}{A single cell ExpressionSet object. This dataset is modified
in-place to include metacluster identifiers.}

\item{var_cluster_name}{A string specifying the column name in sc.eset where
the cluster identifiers are stored. Default is 'cluster_normalized'.}

\item{fit_hierarchical}{A named vector where the names correspond to
celltypes and the values represent the hierarchical clustering fit for each
celltype.}
}
\value{
An ExpressionSet object (sc.eset) with additional metacluster
information.
}
\description{
This function enhances a single cell dataset by adding metacluster
information to it. It achieves this by iterating through each unique value
in the hierarchical fit, identifying the celltypes associated with that
value, and tagging the corresponding cells in the single cell dataset with a
new metacluster identifier.
}
