% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{add_grouping_single_cell}
\alias{add_grouping_single_cell}
\title{Add Grouping to Single Cell Data}
\usage{
add_grouping_single_cell(
  single_cell_data_exp,
  sub_clusters_var,
  level_group_name,
  hierarchy_level_groups
)
}
\arguments{
\item{single_cell_data_exp}{A list containing the single cell data. The
function will add new labels to this list.}

\item{sub_clusters_var}{A string representing the variable used to indicate
the subclusters in \code{single_cell_data_exp}.}

\item{level_group_name}{A string representing the name of the level group
variable in \code{single_cell_data_exp}. The function will add labels to this
variable.}

\item{hierarchy_level_groups}{A list of cell types that define the hierarchy
levels. Each cell type should be a list itself with the elements
\code{celltype_list}, \code{tree_level}, and \code{leaf_number}.}
}
\value{
The function returns the \code{single_cell_data_exp} list with added
labels to the \code{level_group_name} variable.
}
\description{
The \code{add_grouping_single_cell} function iterates over a list of cell types
and assigns labels to each single cell in a dataset based on its group at
each level. The labels are in the format 'subcluster_level_#\emph{leaf}#'.
}
