% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{aggregate_true_top_clustering_prop}
\alias{aggregate_true_top_clustering_prop}
\title{Aggregate True Proportions for Top-Level Clustering}
\usage{
aggregate_true_top_clustering_prop(
  single_cell_data_exp,
  true_prop,
  top_clusters_var,
  top_clusters_list,
  sub_clusters_var
)
}
\arguments{
\item{single_cell_data_exp}{A data frame or matrix containing the single
cell data.}

\item{true_prop}{A matrix or data frame containing the true proportions of
the subclusters.}

\item{top_clusters_var}{The variable name of the top clusters in the
\code{single_cell_data_exp}.}

\item{top_clusters_list}{A list containing the identifiers of the top
clusters.}

\item{sub_clusters_var}{The variable name of the subclusters in the
\code{single_cell_data_exp}.}
}
\value{
A matrix that contains the aggregated true proportions of the top
clusters.
}
\description{
\code{aggregate_true_top_clustering_prop} function calculates the true proportions
of top-level clustersbased on the true subclustering proportions.
}
\details{
The function iterates over the top-level clusters in \code{top_clusters_list}.
For each top-level cluster, it identifies the corresponding subclusters and
calculates the sum of their true proportions.
The process continues until the aggregated true proportions for all top-level
clusters have been calculated.
}
