% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{calculate_joint_cor}
\alias{calculate_joint_cor}
\title{Calculate Joint Correlation}
\usage{
calculate_joint_cor(reference, markers)
}
\arguments{
\item{reference}{A matrix or dataframe of reference data, with each
column representing a sample and each row corresponding to a gene.}

\item{markers}{A vector of marker genes, which will be used to subset the
reference dataset for the correlation computation.

@import stats}
}
\value{
A single numeric value representing the average joint correlation.
}
\description{
This function computes the correlation among distinct groups within the
reference data based on a set of marker genes. The function then calculates
the average joint correlation, which reflects the average correlation
coefficient across all possible pairs of different groups.
}
\details{
The function works by initially calculating the correlation matrix of the
reference data subsetted by the specified marker genes.
It then computes the joint correlation by summing all correlation
coefficients, excluding the self-correlations (i.e., correlation of a group
with itself).
This sum is divided by the number of unique pair combinations of different
groups, yielding the average joint correlation.
The result is a single numeric value representing the average degree of
correlation between the different groups in the reference dataset based on
the specified marker genes.
}
