% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_tree_discovery.R
\name{calculate_k_clustering}
\alias{calculate_k_clustering}
\title{Optimal k Clustering for Hierarchical Clustering}
\usage{
calculate_k_clustering(
  single_cell_data_exp,
  reference,
  var_cell_type,
  cell_types,
  k_value,
  var_sample,
  distance_method = "euclidean",
  hclust_method = "average",
  extra_title = NULL,
  verbose = TRUE
)
}
\arguments{
\item{single_cell_data_exp}{A data frame or matrix containing single-cell
expression data.}

\item{reference}{A reference data set or object to guide the clustering
process.}

\item{var_cell_type}{A string or variable indicating the cell type for each
cell.}

\item{cell_types}{A vector of cell types present in the data.}

\item{k_value}{An integer indicating the number of clusters to be created.}

\item{var_sample}{A string or variable indicating the sample for each cell.}

\item{distance_method}{A string indicating the distance measure to be used
for clustering. This must be one of "euclidean", "maximum", "manhattan",
"canberra", "binary" or "minkowski". Any unambiguous substring can be given.
Default is 'euclidean'.}

\item{hclust_method}{A string indicating the method to be used for
hierarchical clustering. This should be (an unambiguous abbreviation of)
one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA),
"mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). Default
is 'average'.}

\item{extra_title}{A string used for the additional title in the plotting.}

\item{verbose}{A logical value indicating whether to print additional output
during the computation. Default is TRUE.}
}
\value{
A list containing the optimal k clustering results.
}
\description{
This function calculates the optimal number (k) of clusters for a
hierarchical clustering given a specific set of cell types. It can be
customized with different distance measures and agglomeration methods.
}
