% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{calculate_markers_level_intersection}
\alias{calculate_markers_level_intersection}
\title{Calculation of Marker Genes Intersection at Different Levels}
\usage{
calculate_markers_level_intersection(
  single_cell_data_exp,
  top_clusters_var,
  sub_clusters_var,
  top_clusters_list,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  marker_strategy = NULL,
  minimum_markers = 4,
  min_delta_cor_threshold = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{single_cell_data_exp}{A list or data.frame representing the single cell
data.}

\item{top_clusters_var}{A string representing the variable in
\code{single_cell_data_exp} used to indicate the top clusters.}

\item{sub_clusters_var}{A string representing the variable in
\code{single_cell_data_exp} used to indicate the subclusters.}

\item{top_clusters_list}{A list or vector indicating the top clusters to be
considered.}

\item{param.logfc.threshold}{A numeric value representing the log fold change
threshold for marker genes. Default is 2.0.}

\item{param.p_val_adj}{A numeric value representing the adjusted p-value
threshold for marker genes. Default is 0.05.}

\item{marker_strategy}{A character string specifying the overall strategy for
marker gene handling. Default is NULL.}

\item{minimum_markers}{The minimum number of marker genes to be selected,
default is set to 4. This can be adjusted based on the analysis
requirements.}

\item{min_delta_cor_threshold}{Minimum difference between the correlation of
t and t-1 steps. The idea is to skip genes that have small changes in
the correlation. For default it is a 5\%.}

\item{verbose}{A logical value indicating whether to print detailed output
during execution. Default is FALSE.}
}
\value{
A list containing the intersection of marker genes at different
levels.
}
\description{
\code{calculate_markers_level_intersection} computes the intersection of marker
genes at various levels of clustering. This function is currently under
testing and development, intending to streamline the code and improve the
algorithm performance.
}
