% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{calculate_performance_metrics}
\alias{calculate_performance_metrics}
\title{Compute Performance Metrics}
\usage{
calculate_performance_metrics(true_prop, calc_prop, digits = 20)
}
\arguments{
\item{true_prop}{A matrix or dataframe of the true proportions. Rows
correspond to cell types and columns to samples.}

\item{calc_prop}{A matrix or dataframe of the estimated proportions. It
should have the same structure as true_prop.}

\item{digits}{The number of decimal places to which the computed metrics
should be rounded. Default is 20.}
}
\value{
A list of calculated performance metrics. Each element of the list
represents a different metric.
}
\description{
This function calculates several performance metrics to assess the quality
of deconvolution results. These metrics quantify the accuracy of the
estimated cell type proportions compared to the true proportions.
}
\details{
This function computes performance metrics including
Mean Absolute Error (MAE), Root Mean Squared Error (RMSE),
and Pearson Correlation Coefficient (r) to assess the accuracy of
deconvolution results. The metrics are calculated separately for each cell
type and for each sample. The precision of the returned metrics can be
adjusted with the 'digits' parameter.
}
