% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{calculate_prop_back_propagation}
\alias{calculate_prop_back_propagation}
\title{Rescale Clustering Proportions Based on Higher Level Data}
\usage{
calculate_prop_back_propagation(
  top_proportions,
  level_proportions,
  verbose = FALSE
)
}
\arguments{
\item{top_proportions}{A numeric or matrix of proportions representing higher
level clustering.}

\item{level_proportions}{A numeric or matrix of proportions representing more
detailed clustering.}

\item{verbose}{A logical value indicating whether to print detailed output
during execution. Default is FALSE.}
}
\value{
A numeric or matrix of rescaled proportions.
}
\description{
\code{calculate_prop_back_propagation} function computes rescaled proportions of
higher level clustering based on more detailed proportions. It assumes the
input \code{top_proportions} represent proportions for single-leaf samples in the
deconvolution.
}
\details{
The function applies a specific calculation depending on whether
\code{top_proportions} is numeric or a matrix. If it's numeric, it simply
multiplies \code{level_proportions} by \code{top_proportions}.
If it's a matrix, it applies a sweep operation to multiply each column of
\code{level_proportions} by the corresponding column in \code{top_proportions}. The
resulting proportions should sum up to the same values as \code{top_proportions}.
}
