% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_sparseness_matrix}
\alias{calculate_sparseness_matrix}
\title{Calculate Sparseness of a Matrix}
\usage{
calculate_sparseness_matrix(matrix_object)
}
\arguments{
\item{matrix_object}{Matrix, the matrix for which to calculate sparseness.}
}
\value{
Numeric, the sparseness of the matrix in percentage.
}
\description{
This function calculates the sparseness of a given matrix, i.e., the
percentage of zero and NA values in the matrix.
}
\details{
The sparseness of a matrix is defined as the percentage of zero and NA values
out of the total number of entries in the matrix. This function counts the
number of non-zero and non-NA values in the matrix, and then calculates
sparseness as (1 - (non-zero/total)) * 100.
}
