% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{checkHT}
\alias{checkHT}
\title{Check Acceptability of 'n' for 'head()' and 'tail()' Methods}
\usage{
checkHT(n, d)
}
\arguments{
\item{n}{The number to be checked for its acceptability.}

\item{d}{Data input for the function.}
}
\value{
Returns TRUE if 'n' is acceptable, and FALSE otherwise.
}
\description{
This function checks the acceptability of 'n' for 'head()' and
'tail()' methods. If you encounter an error related to this function, you
can load it from:
https://svn.r-project.org/R/trunk/src/library/utils/R/head.R
}
