% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{create_basis_one}
\alias{create_basis_one}
\title{create_basis_one}
\usage{
create_basis_one(
  x,
  ct.sub = NULL,
  ct.varname,
  sample,
  ct.cell.size = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{An ExpressionSet object for single cells. This object will serve as
the basis for the construction of the basis matrix.}

\item{ct.sub}{A vector of cell types that are selected to construct the basis
matrix. This parameter is optional, with the default being NULL.}

\item{ct.varname}{A string that indicates the variable name for 'cell types'
in the 'x' dataset.}

\item{sample}{A string that indicates the variable name for subject/samples
in the 'x' dataset.}

\item{ct.cell.size}{A vector of cell size factors corresponding to the ct.sub
according to prior knowledge. The default is NULL, which means the "library
size" is calculated based on the data. If a vector is provided, the names
of the vector (names(ct.cell.size)) should not be NULL.}

\item{verbose}{A boolean flag for controlling the amount of information
printed during the execution of the function. Default is FALSE.}
}
\value{
A list with the following components:
\itemize{
\item Basis matrix
\item Sum of cell-type-specific library size
\item Sample variance matrix
\item Basis matrix by mvw
\item mvw matrix.
}
}
\description{
This function constructs a basis matrix for single cells from one subject,
based on an ExpressionSet object. It is based on the function created in the
SCDC R implementation. Please refer to the original implementation for more
information: https://meichendong.github.io/SCDC/. Version: 0.0.0.9000
}
\references{
SCDC: a proportion estimation method for single cell RNA-Seq data.
\url{https://meichendong.github.io/SCDC/}.
}
