% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dataframe_results_multiple_deconvolutions}
\alias{create_dataframe_results_multiple_deconvolutions}
\title{Create Data Frame of Multiple Deconvolution Results}
\usage{
create_dataframe_results_multiple_deconvolutions(grid_info, verbose = TRUE)
}
\arguments{
\item{grid_info}{A list of deconvolution results obtained from a grid search.}

\item{verbose}{Shows details about the added deconvolutions.}
}
\value{
A data frame that contains performance metrics for each deconvolution result,
including alpha, beta, divergence value, delta divergence value, and their
logarithmic forms.
}
\description{
This function creates a summary data frame from the results of multiple
deconvolutions, typically obtained from a grid search procedure.
}
\details{
The function iterates through all the deconvolution results in the input list.
For each result, it extracts relevant performance metrics such as alpha, beta,
divergence value, and delta divergence value. These metrics are then combined
into a single data frame. Additionally, it computes the logarithm of the
divergence and delta divergence values, and includes them in the data frame
as well.
}
