% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMMFlex_wrapper.R
\name{environment_creation}
\alias{environment_creation}
\title{Set up environment for NMMFlex}
\usage{
environment_creation()
}
\value{
This function does not return any value.
}
\description{
The environment_creation function installs and verifies the NMMFlex Python
package to prepare the required environment for using NMMFlex in R.
}
