% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_tree_discovery.R
\name{filter_main_celltypes}
\alias{filter_main_celltypes}
\title{Filter Main Cell Types with Markers}
\usage{
filter_main_celltypes(
  single_cell_data_exp,
  reference_celltypes_top,
  group_clusters_var,
  use_min_cor_strategy = TRUE,
  delete_shared_level_markers = FALSE,
  delete_shared_internal_markers = FALSE,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{single_cell_data_exp}{A data frame or matrix containing single-cell
expression data.}

\item{reference_celltypes_top}{A list of reference cell types to be used for
marker identification.}

\item{group_clusters_var}{A string or variable indicating the clusters for
each cell.}

\item{use_min_cor_strategy}{A logical value indicating whether to use the
minimum correlation strategy. Default is TRUE.}

\item{delete_shared_level_markers}{A logical value indicating whether to
delete markers shared at level of the hierarchy. Default is FALSE.}

\item{delete_shared_internal_markers}{A logical value indicating whether to
delete markers shared internally. Default is FALSE.}

\item{param.logfc.threshold}{A numeric value indicating the log-fold-change
threshold for marker selection. Default is 2.0.}

\item{param.p_val_adj}{A numeric value indicating the adjusted p-value
threshold for marker selection. Default is 0.05.}

\item{verbose}{A logical value indicating whether to print additional output
during the computation. Default is FALSE.}
}
\value{
A data frame or matrix containing the filtered single-cell data.
}
\description{
Filters the main cell types using marker genes and performs clustering
optimization on the filtered data.
}
