% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{FindMarkers.proxy}
\alias{FindMarkers.proxy}
\title{Calculate Marker Genes Proxy}
\usage{
\method{FindMarkers}{proxy}(
  seurat.reference,
  group,
  rest_groups,
  test.use.value = "wilcox",
  marker_strategy = "keep_default_values",
  max.size.percentage = 0.7,
  param.logfc.threshold = 2,
  verbose = TRUE
)
}
\arguments{
\item{seurat.reference}{A Seurat object used as the reference for marker gene
calculation. This object contains the single-cell RNA-seq data.}

\item{group}{A string specifying the cell group (based on clustering) for
which to calculate marker genes.}

\item{rest_groups}{A list or vector specifying the remaining cell groups to
be considered as background in the marker genes calculation.}

\item{test.use.value}{String specifying the statistical test to use for
marker identification. The methods currently supported include:
'wilcox', 'bimod', 't', 'negbinom', 'poisson', 'LR', and 'MAST'. Please note
that 'roc' and 'DESeq2' methods are currently not supported due to
discrepancies in output values and support issues, respectively. Default is
'wilcox'.}

\item{marker_strategy}{A string specifying the strategy to use in marker
gene calculation.
'keep_default_values' (default) retains the default values. Other strategies
may be added in the future.}

\item{max.size.percentage}{Maximum acceptable size (as a proportion) for a
group in the marker gene calculation. Default is 0.70.}

\item{param.logfc.threshold}{Threshold for the log-fold change, which
determines the minimum change required for a gene to be considered as a
marker. Default is 2.0.}

\item{verbose}{Logical value indicating whether to display detailed progress
messages during function execution. Default is FALSE.}
}
\value{
A list of marker genes for the specified cell group.
}
\description{
The \code{FindMarkers.proxy} function is a flexible interface for identifying
marker genes using various methods.It's designed to be readily extensible,
facilitating the integration of novel marker calculation techniques in the
future. The function calculates the markers for a specific group relative to
all other cells (rest_groups) in a given Seurat object (seurat.reference).
}
\details{
This function serves as a proxy to different methods for marker gene
calculation. It currently supports several statistical tests and provides a
convenient point of extension for the inclusion of additional marker
calculation methods. Note that the support for different tests can be
influenced by the characteristics of the single-cell experiments, and it's
important to choose a method that's appropriate for your data.
}
