% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources.R
\name{generateBulk_allcells}
\alias{generateBulk_allcells}
\title{generateBulk_allcells}
\usage{
generateBulk_allcells(
  eset,
  ct.varname,
  sample,
  disease = NULL,
  ct.sub = NULL,
  verbose = FALSE
)
}
\arguments{
\item{eset}{An ExpressionSet object for single cells. This object will serve
as the basis for the construction of the Pseudo bulk samples.}

\item{ct.varname}{A string that indicates the variable name for 'cell types'
in the 'eset' dataset.}

\item{sample}{A string that indicates the variable name for subject/samples
in the 'eset' dataset.}

\item{disease}{A string indicating the health condition of subjects.}

\item{ct.sub}{A vector of cell types that are selected to construct Pseudo
bulk samples. This parameter is optional, with the default being NULL. If
NULL, then all cell types are used.}

\item{verbose}{A boolean flag for controlling the amount of information
printed during the execution of the function. Default is FALSE.}
}
\value{
A list with the following components:
\itemize{
\item Pseudo bulk samples ExpressionSet
\item Actual cell-type proportions
}
}
\description{
This function constructs Pseudo bulk samples for single cells from one
subject based on an ExpressionSet object. It is based on the function
created in the SCDC R implementation. Please refer to the original
implementation for more information: https://meichendong.github.io/SCDC/.
Version: 0.0.0.9000
}
\references{
SCDC: a proportion estimation method for single cell RNA-Seq data.
\url{https://meichendong.github.io/SCDC/}.
}
