% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{join_back_propagation_proportions_leaves}
\alias{join_back_propagation_proportions_leaves}
\title{Join Back-Propagation Proportions from Multiple Deconvolution Results}
\usage{
join_back_propagation_proportions_leaves(deco_results_list, verbose = FALSE)
}
\arguments{
\item{deco_results_list}{A list containing deconvolution results. Each
element in the list shouldbe a list with an element named
"back_propagation_proportions_top_detailed" that contains the
back-propagated proportions.}

\item{verbose}{A logical value indicating whether to print detailed output
during execution. Default is FALSE.}
}
\value{
A matrix that contains the merged back-propagation proportions from
all the input deconvolution results.
}
\description{
\code{join_back_propagation_proportions_leaves} function iterates over a list of
deconvolution results and merges the back-propagated proportions from each
result.
}
\details{
The function iterates over the deconvolution results in \code{deco_results_list}.
From each result, it extracts the back-propagation proportions, then merges
them row-wise using the \code{rbind} function.
This process continues until all the back-propagation proportions from all
deconvolution results have been merged.
}
