% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMMFlex_wrapper.R
\name{load_libraries}
\alias{load_libraries}
\title{Load Required Libraries}
\usage{
load_libraries()
}
\value{
A message indicating whether or not the NMMFlex library was
successfully loaded.
}
\description{
This function is designed to load the NMMFlex library, a Python package.
The function uses the \code{reticulate} package to facilitate interoperability
between R and Python, allowing the use of Python scripts within the R
environment. NMMFlex needs to be located at the specific path on your local
machine for this function to work.
}
\details{
This function first attempts to install the Python package from a local
directory using \code{reticulate::py_install}. It then tries to import the
package. If the import is successful, it confirms that the package is
installed and loaded. If not, it throws an error message. Make sure to
adjust the path to the location of the NMMFlex package on your machine.
}
\note{
Ensure that the Python package is located at the provided path on your
machine before using this function.
}
\seealso{
\code{\link{reticulate}}
}
