% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_result_generic}
\alias{order_result_generic}
\title{Order Columns in a Matrix}
\usage{
order_result_generic(matrix, start = 7, verbose = FALSE)
}
\arguments{
\item{matrix}{Matrix, the matrix to be ordered.}

\item{start}{Integer, the column from which to start ordering (default = 7).}

\item{verbose}{Logical, if TRUE, displays detailed messages during execution
(default = FALSE).}
}
\value{
Matrix, the ordered matrix.
}
\description{
This function orders columns in a matrix based on a given starting position.
}
\details{
The function orders the columns in the matrix starting from the position
indicated by the 'start' parameter. It creates a new matrix by concatenating
the ordered columns and returns it.
}
