% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_simple_vector}
\alias{order_simple_vector}
\title{Order Elements in a Vector}
\usage{
order_simple_vector(vector, start = 5, verbose = FALSE)
}
\arguments{
\item{vector}{Vector, the vector to be ordered.}

\item{start}{Integer, the position from which to start ordering
(default = 5).}

\item{verbose}{Logical, if TRUE, displays detailed messages during execution
(default = FALSE).}
}
\value{
Vector, the ordered vector.
}
\description{
This function orders elements in a vector based on a given starting position.
}
\details{
The function orders the elements in the vector starting from the position
indicated by the 'start' parameter. It creates a new vector by concatenating
the ordered elements and returns it.
}
