% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_deconvolution_performance}
\alias{plot_deconvolution_performance}
\title{Plot Deconvolution Performance Metrics}
\usage{
plot_deconvolution_performance(running_info, title_addition = "")
}
\arguments{
\item{running_info}{A data frame with iteration number and the corresponding
values for divergence and delta divergence.}

\item{title_addition}{An optional string that is added to the plot title.}
}
\value{
The function returns a list of ggplot objects, each representing a different
visualization of the deconvolution performance.
}
\description{
This function plots various performance metrics for a deconvolution process
across multiple iterations. Function for gridSearch analysis: General plots.
}
\details{
This function creates three plots to visualize the performance of a
deconvolution process:
\enumerate{
\item The natural logarithm of the divergence value over iterations.
\item The natural logarithm of the delta divergence value over iterations.
\item Both the above metrics in the same plot.
The divergence value provides a measure of how much the deconvolution results
deviate from the ground truth, while the delta divergence gives a measure of
how much the divergence changes between successive iterations.
}
}
