% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_tree_discovery.R
\name{plot_heatmap_correlation_reference}
\alias{plot_heatmap_correlation_reference}
\title{Plot Heatmap of Correlation Among Cell-Types or Groups}
\usage{
plot_heatmap_correlation_reference(reference_w.k.celltypes, title = "Heatmap")
}
\arguments{
\item{reference_w.k.celltypes}{A data frame containing the cell-types or
groups of cell-types and their corresponding basis values.}

\item{title}{Optional; a character string representing the title of the
heatmap. If not provided, the title defaults to 'Heatmap'.}
}
\value{
This function generates a heatmap as a side-effect. It does not
return any value.

@import grDevices
@import ggplot2
@import gplots
@import graphics
@import stats
}
\description{
This function creates a heatmap to visually display the correlation between
different cell-types or groups of cell-types. The color gradient of the
heatmap represents the correlation coefficient values, providing a clear
visual summary of the correlations in the data.
}
