% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_performance_for_each_model}
\alias{plot_performance_for_each_model}
\title{Plot Performance for Each Model}
\usage{
plot_performance_for_each_model(
  grid_search_result,
  alpha = NULL,
  beta = NULL,
  plot_type = "summary",
  scale_parameters = FALSE,
  save.pdf = FALSE,
  path.pdf = NULL
)
}
\arguments{
\item{grid_search_result}{A list object that contains the results of a grid
search.}

\item{alpha}{Numeric, the alpha value to filter the models to plot. Optional,
default is NULL.}

\item{beta}{Numeric, the beta value to filter the models to plot. Optional,
default is NULL.}

\item{plot_type}{Character, specifies the type of plot to produce. Options
are 'summary' (default) or other types supported by the underlying plotting
function.}

\item{scale_parameters}{Logical, whether to scale the parameters in the plot.
Optional, default is FALSE.}

\item{save.pdf}{Logical, whether to save the plot as a PDF. Optional, default
is FALSE.}

\item{path.pdf}{String, the path where to save the PDF file if 'save.pdf' is
TRUE. Optional, default is NULL.}
}
\value{
The function is used for its side effects of creating and optionally saving
plots. It does not return anything.
}
\description{
This function plots the performance of each model from a grid search result
based on specified alpha and beta values.
}
\details{
The function iterates over each model in the grid search result. For each
model, it produces a plot based on the 'plot_type' argument.
It filters the models to plot based on the 'alpha' and 'beta' arguments.

If 'save.pdf' is TRUE, it saves the plot as a PDF file to the specified path
or the current working directory if no path is specified.
}
