% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{print_performance_metrics}
\alias{print_performance_metrics}
\title{Display Performance Metrics}
\usage{
print_performance_metrics(list_metrics, title = NULL)
}
\arguments{
\item{list_metrics}{A list of calculated performance metrics, as produced
by the \code{calculate_performance_metrics()} function.}

\item{title}{An optional title for the printed metrics. Default is NULL.}
}
\value{
NULL. The function is used for its side effect of printing the
metrics.
}
\description{
This function neatly prints the calculated performance metrics, facilitating
the interpretation of deconvolution results.
}
\details{
The function takes as input a list of performance metrics, which should
include metrics such as Mean Absolute Error (MAE),
Root Mean Squared Error (RMSE), and Pearson Correlation Coefficient (r).
These metrics are then printed in a structured format for easy
interpretation. The optional 'title' parameter can be used to give the
printed metrics a descriptive title.
}
