% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{run_deconvolution_simulation}
\alias{run_deconvolution_simulation}
\title{Deconvolution Simulation}
\usage{
run_deconvolution_simulation(
  single_cell_data_exp,
  top_clusters_var,
  top_clusters_list,
  sub_clusters_var,
  sub_clusters_list,
  sample,
  use_min_cor_strategy = FALSE,
  delete_shared_level_markers = TRUE,
  delete_shared_internal_markers = TRUE,
  filter_markers = NULL,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  marker_strategy = "keep_default_values",
  verbose = FALSE
)
}
\arguments{
\item{single_cell_data_exp}{A data frame representing the single cell data
for the deconvolution simulation.}

\item{top_clusters_var}{A vector indicating the top clusters variables.}

\item{top_clusters_list}{A list containing the top clusters.}

\item{sub_clusters_var}{A vector indicating the sub clusters variables.}

\item{sub_clusters_list}{A list containing the sub clusters.}

\item{sample}{A numeric or integer value indicating the sample size for the
deconvolution simulation.}

\item{use_min_cor_strategy}{A boolean. If TRUE, the function uses minimum
correlation strategy. Default is FALSE.}

\item{delete_shared_level_markers}{A boolean. If TRUE, shared level markers
will be deleted. Default is TRUE.}

\item{delete_shared_internal_markers}{A boolean. If TRUE, shared internal
markers will be deleted. Default is TRUE.}

\item{filter_markers}{A vector of markers to filter, or NULL. Default is NULL}

\item{param.logfc.threshold}{A numeric value indicating the log fold change
threshold. Default is 2.0.}

\item{param.p_val_adj}{A numeric value indicating the adjusted p-value
threshold. Default is 0.05.}

\item{marker_strategy}{A character string specifying the marker strategy.
Default is 'keep_default_values'.}

\item{verbose}{A boolean. If TRUE, the function will print additional details
during the execution. Default is FALSE.}
}
\value{
A list containing the results of the deconvolution simulation. This
might include the estimated sources, the final weights, and any other
relevant output depending on the actual implementation of the function.
}
\description{
This function performs a detailed simulation of the
deconvolution process on single cell expression data. It allows for the
inclusion of multiple cluster levels (both top level and sub level clusters),
and customization of various parameters such as marker strategy, log fold
change threshold, and adjusted p-value threshold. The function also supports
a minimum correlation strategy for deconvolution, providing flexibility in
the simulation approach. Verbose mode can be enabled for additional execution
details.
Note: If you encounter an error related to the function 'checkHT', you can
load it from: https://svn.r-project.org/R/trunk/src/library/utils/R/head.R
}
\details{
Tree Guided Deconvolution Functions

This script, tree_guided_deconvolution.R, is part of the Decoflex package.
It includes the following functions:

\describe{
\item{add_grouping_single_cell()}{A function for adding single cell
groupings.}
\item{aggregate_true_top_clustering_prop()}{A function to aggregate true
top clustering proportions.}
\item{calculate_joint_cor()}{A function to calculate joint correlation.}
\item{calculate_markers()}{A function to calculate markers.}
\item{calculate_markers_level_intersection()}{A function to calculate
markers level intersection.}
\item{calculate_min_correlation_incremental_markers()}{A function to
calculate minimum correlation incremental markers.}
\item{calculate_performance_metrics()}{A function to calculate performance
metrics.}
\item{calculate_prop_back_propagation()}{A function to calculate proportion
back propagation.}
\item{FindMarkers.proxy()}{A proxy function to find markers.}
\item{join_back_propagation_proportions_leaves()}{A function to join back
propagation proportions leaves.}
\item{performance_metrics_general()}{A function to calculate general
performance metrics.}
\item{print_performance_metrics()}{A function to print performance
metrics.}
\item{recalculate_proportions()}{A function to recalculate proportions.}
\item{run_deconvolution_simulation_generic()}{A function to run generic
deconvolution simulations.}
\item{run_deconvolution_tree_guided()}{A function to run tree-guided
deconvolution.}
\item{run_deconvolution_tree_guided_recursive()}{A recursive function to
run tree-guided deconvolution.}
}
}
