% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{run_deconvolution_simulation_generic}
\alias{run_deconvolution_simulation_generic}
\title{Generic Deconvolution Simulation}
\usage{
run_deconvolution_simulation_generic(
  single_cell_data_exp,
  top_clusters_var,
  top_clusters_list,
  sub_clusters_var,
  sub_clusters_list,
  sample,
  use_min_cor_strategy = FALSE,
  delete_shared_level_markers = TRUE,
  delete_shared_internal_markers = TRUE,
  filter_markers = NULL,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  marker_strategy = NULL,
  verbose = FALSE
)
}
\arguments{
\item{single_cell_data_exp}{A data frame representing the single cell data
for the deconvolution simulation.}

\item{top_clusters_var}{A vector indicating the top clusters variables.}

\item{top_clusters_list}{A list containing the top clusters.}

\item{sub_clusters_var}{A vector indicating the sub clusters variables.}

\item{sub_clusters_list}{A list containing the sub clusters.}

\item{sample}{A numeric or integer value indicating the sample size for the
deconvolution simulation.}

\item{use_min_cor_strategy}{A boolean. If TRUE, the function uses minimum
correlation strategy. Default is FALSE.}

\item{delete_shared_level_markers}{A boolean. If TRUE, shared level markers
will be deleted. Default is TRUE.}

\item{delete_shared_internal_markers}{A boolean. If TRUE, shared internal
markers will be deleted. Default is TRUE.}

\item{filter_markers}{A vector of markers to filter, or NULL. Default is NULL}

\item{param.logfc.threshold}{A numeric value indicating the log fold change
threshold. Default is 2.0.}

\item{param.p_val_adj}{A numeric value indicating the adjusted p-value
threshold. Default is 0.05.}

\item{marker_strategy}{A character string specifying the marker strategy, or
NULL. Default is NULL.}

\item{verbose}{A boolean. If TRUE, the function will print additional details
during the execution. Default is FALSE.}
}
\value{
A list containing the results of the deconvolution simulation. This
might include the estimated sources,
the final weights, and any other relevant output depending on the actual
implementation of the function.
}
\description{
This function is a generalized form of the deconvolution
simulation process applicable to single cell data. It allows users to modify
various parameters like the marker strategy, log fold change threshold, and
adjusted p-value threshold. By setting use_min_cor_strategy to TRUE, the
function will apply a minimum correlation strategy during the deconvolution.
}
