% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{run_deconvolution_simulation_generic_recursive}
\alias{run_deconvolution_simulation_generic_recursive}
\title{run_deconvolution_simulation_generic_recursive}
\usage{
run_deconvolution_simulation_generic_recursive(
  single_cell_data_exp,
  hierarchy,
  sub_clusters_var,
  sample,
  use_min_cor_strategy = TRUE,
  delete_shared_level_markers = FALSE,
  delete_shared_internal_markers = FALSE,
  deconvolute_just_top = FALSE,
  filter_markers = NULL,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  marker_strategy = NULL,
  verbose = FALSE
)
}
\arguments{
\item{single_cell_data_exp}{A data frame that contains single cell expression
data.}

\item{hierarchy}{A list that outlines the hierarchy of cell types.}

\item{sub_clusters_var}{A variable that represents the subclusters in the
data.}

\item{sample}{A variable that represents the samples in the data.}

\item{use_min_cor_strategy}{A logical parameter indicating whether to use
minimum correlation strategy. Defaults to TRUE.}

\item{delete_shared_level_markers}{A logical parameter indicating whether to
delete shared level markers. Defaults to FALSE.}

\item{delete_shared_internal_markers}{A logical parameter indicating whether
to delete shared internal markers. Defaults to FALSE.}

\item{deconvolute_just_top}{A logical parameter indicating whether to only
deconvolute top. Defaults to FALSE.}

\item{filter_markers}{A parameter for specifying markers to filter, defaults
to NULL.}

\item{param.logfc.threshold}{A numeric value specifying the threshold for
log fold change, defaults to 2.0.}

\item{param.p_val_adj}{A numeric value specifying the threshold for adjusted
p-values, defaults to 0.05.}

\item{marker_strategy}{Strategy for selecting markers, defaults to NULL.}

\item{verbose}{A logical parameter indicating whether to print detailed
messages during the execution of the function. Defaults to FALSE.}
}
\value{
A list containing the deconvolution result and the pseudo bulk data.
}
\description{
Perform a recursive deconvolution simulation on given single
cell data. This function handles the first top node of the hierarchy and then
send the rest to the recursive algorithm
}
