% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_guided_deconvolution.R
\name{run_deconvolution_tree_guided_recursive}
\alias{run_deconvolution_tree_guided_recursive}
\title{Run Deconvolution Tree Guided Recursively}
\usage{
run_deconvolution_tree_guided_recursive(
  result_deco_top = NULL,
  bulk_data,
  true_proportions = NULL,
  single_cell_data_exp,
  sub_clusters_var,
  hierarchy,
  sample,
  use_min_cor_strategy = TRUE,
  delete_shared_level_markers = FALSE,
  delete_shared_internal_markers = FALSE,
  deconvolute_just_top = FALSE,
  deconvolute_top_hierarchy_limit = 1,
  filter_markers = NULL,
  param.logfc.threshold = 2,
  param.p_val_adj = 0.05,
  test.use.value = "wilcox",
  marker_strategy = "keep_default_values",
  minimum_markers = 4,
  min_delta_cor_threshold = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{result_deco_top}{A list representing the result of deconvolution at
the top level of the hierarchy. If NULL (default), the function
starts at the root of the tree.}

\item{bulk_data}{The bulk data on which the deconvolution needs to be
performed.}

\item{true_proportions}{An optional matrix representing the true proportions
of cell types. This can be used to calculate the accuracy of
the deconvolution. If NULL (default), the accuracy calculation is skipped.}

\item{single_cell_data_exp}{A data frame representing single cell expression
data, used to guide the deconvolution process.}

\item{sub_clusters_var}{A string representing the variable in
single_cell_data_exp which indicates sub-clusters.}

\item{hierarchy}{A list structure describing the hierarchy of cell type
clustering.}

\item{sample}{A string representing the sample name.}

\item{use_min_cor_strategy}{A boolean indicating whether to use the minimum
correlation strategy during the marker calculation phase. Default is TRUE.}

\item{delete_shared_level_markers}{A boolean indicating whether to delete
shared level markers during the marker calculation phase.
Default is FALSE.}

\item{delete_shared_internal_markers}{A boolean indicating whether to delete
shared internal markers during the marker calculation phase.
Default is FALSE.}

\item{deconvolute_just_top}{A boolean indicating whether to perform
deconvolution only at the top level. If TRUE (default is FALSE), the
function stops after deconvoluting the top level and returns the result.}

\item{deconvolute_top_hierarchy_limit}{An integer indicating the top
hierarchy limit for deconvolution. Default is 1.}

\item{filter_markers}{An optional list of markers to be used for filtering
during the marker calculation phase. If NULL (default), no filtering is
performed.}

\item{param.logfc.threshold}{A numeric value representing the log-fold-change
threshold for marker calculation. Default is 2.0.}

\item{param.p_val_adj}{A numeric value representing the adjusted p-value
threshold for marker calculation. Default is 0.05.}

\item{test.use.value}{A string indicating the statistical test to use for
marker calculation. Default is 'wilcox'.}

\item{marker_strategy}{An optional string representing the marker selection
strategy. If NULL (default), a default strategy is used.}

\item{minimum_markers}{The minimum number of marker genes to be selected,
default is set to 4. This can be adjusted based on the analysis
requirements.}

\item{min_delta_cor_threshold}{Minimum difference between the correlation of
t and t-1 steps. The idea is to skip genes that have small changes in
the correlation. For default it is a 5\%.}

\item{verbose}{A boolean indicating whether to print detailed messages during
the execution of the function. Default is FALSE.}
}
\value{
A list containing the result of the top level deconvolution, a list
of deconvolution results at each level, and the back-propagated proportions
at the top level.
}
\description{
This function uses a tree-guided method to perform a recursive deconvolution
operation on bulk and single-cell data. The recursive method helps in
analyzing hierarchically structured data, usually encountered in biological
systems where a set of cell types are hierarchically organized based on
their molecular similarity. The method starts at the top level,
identifies clusters, calculates markers, performs deconvolution, and
propagates the information back to guide the deconvolution at the next level.
}
