% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMMFlex_wrapper.R
\name{run_grid_search}
\alias{run_grid_search}
\title{Run Grid Search}
\usage{
run_grid_search(
  bulk_data_methylation,
  bulk_data_expression = NULL,
  data_expression_auxiliary = NULL,
  k,
  alpha_list = NULL,
  beta_list = NULL,
  delta_threshold = 1e-20,
  max_iterations = 200,
  print_limit = 100,
  threads = 0,
  proportion_constraint_h = TRUE,
  regularize_w = NULL,
  alpha_regularizer_w_list = NULL,
  fixed_w = NULL,
  fixed_h = NULL,
  fixed_a = NULL,
  fixed_b = NULL
)
}
\arguments{
\item{bulk_data_methylation}{A data.frame, Bulk data methylation matrix.}

\item{bulk_data_expression}{A data.frame, Bulk data expression matrix.}

\item{data_expression_auxiliary}{A data.frame, Auxiliary expression data
matrix.}

\item{k}{An integer, Number of clusters.}

\item{alpha_list}{A list of numerics, List of alpha values to be considered
in the grid search. If NULL, no grid search is performed over alpha values.}

\item{beta_list}{A list of numerics, List of beta values to be considered in
the grid search. If NULL, no grid search is performed over beta values.}

\item{delta_threshold}{A numeric, The threshold value for convergence.
Default is 1e-20.}

\item{max_iterations}{An integer, Maximum number of iterations for
convergence. Default is 200.}

\item{print_limit}{An integer, Limit for print statements. Default is 100.}

\item{threads}{An integer, Number of CPU threads to be used. If 0, then it
uses the total number of CPUs minus one. Default is 0.}

\item{proportion_constraint_h}{A logical, Whether to apply proportion
constraint on H matrix. Default is TRUE.}

\item{regularize_w}{A logical, Whether to apply regularization on W matrix.
If NULL, no regularization is applied.}

\item{alpha_regularizer_w_list}{A list of numerics, List of alpha regularizer
values to be considered in the grid search. If NULL, no grid search is
performed over alpha regularizer values.}

\item{fixed_w}{A matrix or NULL, Fixed matrix W. If NULL, it implies that W
is not fixed.}

\item{fixed_h}{A matrix or NULL, Fixed matrix H. If NULL, it implies that H
is not fixed.}

\item{fixed_a}{A matrix or NULL, Fixed matrix A. If NULL, it implies that A
is not fixed.}

\item{fixed_b}{A matrix or NULL, Fixed matrix B. If NULL, it implies that B
is not fixed.}
}
\value{
A list of results from the grid search.
}
\description{
Performs a grid search in a parallelized manner over different
values of alpha and beta in the Non-negative matrix factorization. The search
is conducted over the combinations of given alpha and beta values.
}
\details{
Raises an error if gamma, alpha, and beta are all 0, indicating that the
model cannot be executed.
Suggests the user to switch to the more direct function
run_deconvolution_multiple() if any two of gamma, alpha, beta are zero.
}
