% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMMFlex_wrapper.R
\name{run_standard_deconvolution}
\alias{run_standard_deconvolution}
\title{Run standard deconvolution}
\usage{
run_standard_deconvolution(
  bulk_data_x,
  references_w,
  markers = NULL,
  max_iterations = 10000,
  verbose = FALSE
)
}
\arguments{
\item{bulk_data_x}{A numeric vector, matrix, or data frame representing the
bulk data for the deconvolution.}

\item{references_w}{A numeric vector, matrix, or data frame representing the
references for the deconvolution.}

\item{markers}{A numeric vector or NULL. Represents the markers to be used
in the deconvolution. By default, this is set to NULL.}

\item{max_iterations}{A positive integer. Represents the maximum number of
iterations for the deconvolution process. By default, this is set to 10000.}

\item{verbose}{A boolean. If TRUE, the function will print additional details
during the execution. By default, this is set to FALSE.}
}
\value{
A list containing the results of the deconvolution. This might
include the estimated sources and the final weights depending on the
implementation of MMNF library in python.
}
\description{
This function is a wrapper for the MMNF library in python that
is used for running a standard deconvolution process. This involves
estimating the source distribution in a mixture model given some observed
data.
}
