% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot_as_pdf}
\alias{save_plot_as_pdf}
\title{Save Plot as PDF}
\usage{
save_plot_as_pdf(
  plot_object,
  file_name,
  width.parameter = NULL,
  height.parameter = NULL,
  path = NULL
)
}
\arguments{
\item{plot_object}{The plot object to be saved.}

\item{file_name}{String, the name of the file to save the plot as. It should
not include the path or file extension.}

\item{width.parameter}{Numeric, the width of the output PDF in inches.
Optional, default is NULL.}

\item{height.parameter}{Numeric, the height of the output PDF in inches.
Optional, default is NULL.}

\item{path}{String, the directory where the PDF should be saved. Optional,
default is '/mnt_volumen/GIT_REPOSITORIES/MTF/results/'.}
}
\value{
The function is used for its side effects of creating a PDF file and does not
return anything.
}
\description{
This function saves a given plot object as a PDF file to a specific path.
}
\details{
The function first checks whether a path was provided. If not, it uses a
default path. Then, it opens a new PDF device, prints the 'plot_object' to
this device, and then closes the device, which saves the plot to the
specified file. The 'width.parameter' and 'height.parameter' arguments allow
for control over the dimensions of the output PDF. If these are not provided,
the PDF is created with the default dimensions.
}
