% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMMFlex_wrapper.R
\name{trans_grid_search_to_R}
\alias{trans_grid_search_to_R}
\title{Translate Grid Search Results to R}
\usage{
trans_grid_search_to_R(grid_search_result, verbose = FALSE)
}
\arguments{
\item{grid_search_result}{A list. This is the result object returned from a
grid search process.}

\item{verbose}{A logical. If TRUE, the function will print information about
the progress of the transformation for each model in the console. Default
is FALSE.}
}
\value{
A list of all the elements contained in grid_search_result,
transformed into an R-friendly format. Each element in the list corresponds
to a model from the grid search and is named "alpha_beta_" followed by the
alpha and beta parameters of the model.
}
\description{
This function takes the output of a grid search and translates
it into an R-friendly format, returning a list
of all relevant results. It works by iterating over the grid search results,
converting those into native R objects.The idea is to preserve the python
data in the R allocated memory, otherwise under certain circumstance the
object will be null.
}
\details{
This function is designed to simplify the handling of grid search
results by translating the returned objects from Python into an R-friendly
format. The function uses \code{trans_mult_deco_to_R()} to transform each
individual deconvolution result.
}
