% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_deconvolution_performance_x_y_z_total}
\alias{plot_deconvolution_performance_x_y_z_total}
\title{Plot Detailed Deconvolution Performance}
\usage{
plot_deconvolution_performance_x_y_z_total(
  running_info,
  title_addition = "",
  alpha = NULL,
  beta = NULL,
  scale_parameters = FALSE
)
}
\arguments{
\item{running_info}{Data frame, the running information of the deconvolution.}

\item{title_addition}{Character, an optional string to be added to the plot
title. Optional, default is ''.}

\item{alpha}{Numeric, the alpha value to scale the y divergence. Optional,
default is NULL.}

\item{beta}{Numeric, the beta value to scale the z divergence. Optional,
default is NULL.}

\item{scale_parameters}{Logical, whether to scale the y and z divergence
values by the given alpha and beta. Optional, default is FALSE.}
}
\value{
A ggplot2 object representing the plot.
}
\description{
This function plots detailed performance of the deconvolution with respect
to x, y, z, and total divergence. If specified, the y and z divergence values
can be scaled by the given alpha and beta parameters.
}
\details{
The function generates a line plot showing the log of the divergence values
(for x, y, z, total, and delta) against iterations.
The line color is specified according to the type of divergence.
The divergence values for y and z can be scaled by the specified alpha and
beta parameters if 'scale_parameters' is TRUE.
}
